Divide core/paragraph extension block.js into style01.js and border01.js
今後の拡張性を考え、block.jsを分離しました。(今回のようにボーダーが全く同じで良ければborder01.jsにブロック名を追記する)
分離した結果、Buttonとlabelのidに使っていたindex情報を従来のように使い回せなくなったので、他のIDと重複しないよう、
cocoon-dorder-<filename>-button-<index>としました。
