+++ blocks/src/block-extension/style-extension/style01.js
@@ -19,10 +19,6 @@ function addCustomAttributes( settings ) {
         type: 'string',
         default: '',
       },
-      extraBorder: {
-        type: 'string',
-        default: '',
-      },
     } );
   }
   return settings;
@@ -38,7 +34,7 @@ const addCustomEdit = createHigherOrderComponent( ( BlockEdit ) => {
   return ( props ) => {
     if ( allowedBlocks.includes( props.name ) && props.isSelected ) {
       const { setAttributes, isSelected, attributes } = props;
-      const { className, extraStyle, extraBorder } = attributes;
+      const { className, extraStyle } = attributes;
 
       const extraStyles = [
         {
@@ -171,177 +167,12 @@ const addCustomEdit = createHigherOrderComponent( ( BlockEdit ) => {
         },
       ];
 
-      const extraBorders = [
-        {
-          style: '',
-          buttonText: __( 'デフォルト', THEME_NAME ),
-        },
-        {
-          style: 'border-solid',
-          buttonText: __( '実線', THEME_NAME ),
-        },
-        {
-          style: 'border-double',
-          buttonText: __( '二重線', THEME_NAME ),
-        },
-        {
-          style: 'border-dashed',
-          buttonText: __( '破線', THEME_NAME ),
-        },
-        {
-          style: 'border-dotted',
-          buttonText: __( '点線', THEME_NAME ),
-        },
-        {
-          style: 'border-thin-and-thick',
-          buttonText: __( '薄太', THEME_NAME ),
-        },
-        {
-          style: 'border-convex',
-          buttonText: __( '微凸', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-s-solid',
-          buttonText: __( '実線(角丸小)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-s-double',
-          buttonText: __( '二重線(角丸小)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-s-dashed',
-          buttonText: __( '破線(角丸小)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-s-dotted',
-          buttonText: __( '点線(角丸小)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-s-thin-and-thick',
-          buttonText: __( '薄太(角丸小)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-s-convex',
-          buttonText: __( '微凸(角丸小)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-l-solid',
-          buttonText: __( '実線(角丸大)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-l-double',
-          buttonText: __( '二重線(角丸大)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-l-dashed',
-          buttonText: __( '破線(角丸大)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-l-dotted',
-          buttonText: __( '点線(角丸大)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-l-thin-and-thick',
-          buttonText: __( '薄太(角丸大)', THEME_NAME ),
-        },
-        {
-          style: 'border-radius-l-convex',
-          buttonText: __( '微凸(角丸大)', THEME_NAME ),
-        },
-      ];
-
-      var index = 0;
       return (
         <Fragment>
           <BlockEdit { ...props } />
 
           { isSelected && (
             <InspectorControls>
-              <PanelBody
-                title={ __( '[C] ボーダー設定', THEME_NAME ) }
-                initialOpen={ false }
-              >
-                <div class="__clearBtn">
-                  <button
-                    type="button"
-                    class="components-button is-small"
-                    onClick={ () =>
-                      setAttributes( { extraBorder: '' } )
-                    }
-                  >
-                    <span class="dashicons dashicons-editor-removeformatting"></span>
-                    {__( 'ボーダーをクリア', THEME_NAME ) }
-                  </button>
-                </div>
-                <div className="block-editor-block-styles">
-                  <div className="block-editor-block-styles__variants style-buttons">
-                    {extraBorders.map((border) => {
-                      index++;
-                      return (
-                        <div
-                          className={classnames('__btnBox',
-                            {[ '__btnBoxDefault display-none' ]: ! border.style}
-                          )}
-                        >
-                          <Button
-                            id={ 'cocoon-dorder-button-' + index }
-                            className={ classnames(
-                              'display-none',
-                              'block-editor-block-styles__item',
-                              {
-                                'is-active': border.style === extraBorder,
-                              }
-                            ) }
-                            variant="secondary"
-                            label={ border.buttonText }
-                            onClick={ () =>
-                              {
-                                if (border.style === extraBorder) {
-                                  setAttributes( { extraBorder: '' } )
-                                } else {
-                                  setAttributes( { extraBorder: border.style } )
-                                }
-                              }
-                            }
-                          ></Button>
-                          <label
-                            for={ 'cocoon-dorder-button-' + index }
-                            class="__labelBtn"
-                            data-selected={
-                              border.style === extraBorder ? true : false
-                            }
-                          >
-                            <span class="__prevWrap editor-styles-wrapper">
-                              <span
-                                className={ classnames( '__prev', {
-                                  [ 'is-style-' + border.style ]:
-                                    !! border.style,
-                                  [ 'has-border' ]: border.style,
-                                } ) }
-                              ></span>
-                            </span>
-                            <span class="__prevTitle">
-                              { border.buttonText }
-                            </span>
-                          </label>
-                        </div>
-                      );
-                    } ) }
-                  </div>
-                </div>
-                <div class="__clearBtn">
-                  <button
-                    type="button"
-                    class="components-button is-small"
-                    onClick={ () =>
-                      setAttributes( { extraBorder: '' } )
-                    }
-                  >
-                    <span class="dashicons dashicons-editor-removeformatting"></span>
-                    {__( 'ボーダーをクリア', THEME_NAME ) }
-                  </button>
-                </div>
-              </PanelBody>
               <PanelBody
                 title={ __( '[C] スタイル', THEME_NAME ) }
                 initialOpen={ false }
@@ -350,26 +181,23 @@ const addCustomEdit = createHigherOrderComponent( ( BlockEdit ) => {
                   <button
                     type="button"
                     class="components-button is-small"
-                    onClick={ () =>
-                      setAttributes( { extraStyle: '' } )
-                    }
+                    onClick={ () => setAttributes( { extraStyle: '' } ) }
                   >
                     <span class="dashicons dashicons-editor-removeformatting"></span>
-                    {__( 'スタイルをクリア', THEME_NAME ) }
+                    { __( 'スタイルをクリア', THEME_NAME ) }
                   </button>
                 </div>
                 <div className="block-editor-block-styles">
                   <div className="block-editor-block-styles__variants style-buttons">
-                    {extraStyles.map((style) => {
-                      index++
+                    { extraStyles.map( ( style, index ) => {
                       return (
                         <div
-                          className={classnames('__btnBox',
-                            {[ '__btnBoxDefault display-none' ]: ! style.style}
-                          )}
+                          className={ classnames( '__btnBox', {
+                            [ '__btnBoxDefault display-none' ]: ! style.style,
+                          } ) }
                         >
                           <Button
-                            id={ 'cocoon-dorder-button-' + index }
+                            id={ 'cocoon-dorder-style01-button-' + index }
                             className={ classnames(
                               'display-none',
                               'block-editor-block-styles__item',
@@ -379,18 +207,16 @@ const addCustomEdit = createHigherOrderComponent( ( BlockEdit ) => {
                             ) }
                             variant="secondary"
                             label={ style.buttonText }
-                            onClick={ () =>
-                              {
-                                if (style.style === extraStyle) {
-                                  setAttributes( { extraStyle: '' } )
-                                } else {
-                                  setAttributes( { extraStyle: style.style } )
-                                }
+                            onClick={ () => {
+                              if ( style.style === extraStyle ) {
+                                setAttributes( { extraStyle: '' } );
+                              } else {
+                                setAttributes( { extraStyle: style.style } );
                               }
-                            }
+                            } }
                           ></Button>
                           <label
-                            for={ 'cocoon-dorder-button-' + index }
+                            for={ 'cocoon-dorder-style01-button-' + index }
                             class="__labelBtn"
                             data-selected={
                               style.style === extraStyle ? true : false
@@ -417,12 +243,10 @@ const addCustomEdit = createHigherOrderComponent( ( BlockEdit ) => {
                   <button
                     type="button"
                     class="components-button is-small"
-                    onClick={ () =>
-                      setAttributes( { extraStyle: '' } )
-                    }
+                    onClick={ () => setAttributes( { extraStyle: '' } ) }
                   >
                     <span class="dashicons dashicons-editor-removeformatting"></span>
-                    {__( 'スタイルをクリア', THEME_NAME ) }
+                    { __( 'スタイルをクリア', THEME_NAME ) }
                   </button>
                 </div>
               </PanelBody>
@@ -448,15 +272,13 @@ const applyAttributesToBlock = createHigherOrderComponent(
       const { attributes, className, name, isValid } = props;
 
       if ( isValid && allowedBlocks.includes( name ) ) {
-        const { extraStyle, extraBorder } = attributes;
+        const { extraStyle } = attributes;
 
         return (
           <BlockListBlock
             { ...props }
             className={ classnames( className, {
               [ 'is-style-' + extraStyle ]: !! extraStyle,
-              [ 'is-style-' + extraBorder ]: !! extraBorder,
-              [ 'has-border' ]: extraBorder,
               [ 'has-box-style' ]: extraStyle,
             } ) }
           />
@@ -478,12 +300,10 @@ addFilter(
 const addCustomSave = ( props, blockType, attributes ) => {
   if ( allowedBlocks.includes( blockType.name ) ) {
     const { className } = props;
-    const { extraStyle, extraBorder } = attributes;
+    const { extraStyle } = attributes;
 
     props.className = classnames( className, {
       [ 'is-style-' + extraStyle ]: !! extraStyle,
-      [ 'is-style-' + extraBorder ]: !! extraBorder,
-      [ 'has-border' ]: extraBorder,
       [ 'has-box-style' ]: extraStyle,
     } );
