+++ blocks/src/blocks.js
@@ -115,7 +115,8 @@ registerCocoonBlocks();
 
 //デフォルトブロックの拡張
 import './custom/code/block.js';
-import './block-extension/style-extension/block.js';
+import './block-extension/style-extension/border01.js';
+import './block-extension/style-extension/style01.js';
 
 //レイアウト
 import './layout/column-children/block.js';
